Added session summary view.
Changed data types for SessionData attributes from int to float in sql scripts as precision was lost.
Changed StringReader to check for tabbed spacing and not just single spaces.
Added MaxBy and MinBy MoreLinq helpers to return min and max values of an attribute in a collection.
Mocked context class in SessionRepositoryTests class.