Added line graph for SessionData to the summary view.
Added highchart.js to project and the ng-highchart angular directive.
The GetSessionData method in the SessionRepository gets the maximum value from the SessionData table and this value is used to set the Y axis in the graph.
The total count of the SessionData containing the sessionId is set in the SessionDataGraphDto to set the X axis in the graph.

The sessionSummaryCtrl has a 1 second timer before the chart is drawn and summary is displayed.
Added collectionOfObjectsToArray method to convert json response to array. Each array item contains an interval which is incremented so data is plotted along the X axis, and the position to be plotted.
DrawGraph method draws the chart, by calling addSeries, specifying a chart label, line colour and passing in the array of converted data.