Select a portion of the graph and update summary data.
Added FilterObjectDates method which takes a starting date and adds the minimum and maximum amount of seconds passed to it.
All dates that are within range of the minimum and maximum dates are returned.
The method uses DateTime.Compare:
DateTime.Compare returns -1 if date parameter 1 is less than than date parameter 2
DateTime.Compare returns 0 if date parameter 1 is equal to date parameter 2
DateTime.Compare returns 1 if date parameter 1 is greater than date parameter 2

The scope watch functions have been placed in watchMinAndMaxTimes function. This is called in loadSessionDataCompleted so the watchers aren't called too early.
if (newValue !== oldValue) {} checks added so the watchers aren't called when the controller is initialised.