Calculated normalized power and intensity factor.
For session and data subset.

Calculates a rolling 30 second average of the preceding time points after 30 seconds.
Raises all the moving averages to the fourth power.
Finds the average of values raised to fourth power.
Takes the fourth root of the average values raised to the fourth power.

Removed duplicate alerts. Added alert if normalized power and intensity factor cannot be calculated.